working functionality

+++ Backend/src/main/java/laniakea/localgroup/milkyway/sol/earth/gr/hua/dit/ds/assigment/AirTours/controllers/AuthController.java
@@ -58,7 +58,7 @@ public class AuthController {
                 .collect(Collectors.toList());
 
         return ResponseEntity.ok(new JwtResponse(
-                userDetails.getId(),
+                (long) userDetails.getId(),
                 userDetails.getEmail(),
                 userDetails.getFirstName(),
                 userDetails.getLastName(),
