+++ Backend/src/main/java/laniakea/localgroup/milkyway/sol/earth/gr/hua/dit/ds/assigment/AirTours/repository/RoleRepository.java
@@ -7,6 +7,8 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Long > {
+public interface RoleRepository extends JpaRepository<Role, Integer > {
     Optional<Role> findByName(ERole name);
+
+
 }
